requirements.txt
moviepy==1.0.3
matplotlib==3.0.3
scipy==1.3.2
python_vlc==3.0.12118
speechbrain==0.5.6
opencv_python_headless==4.5.2.52
pydub==0.25.1
pandas==0.24.2
face_recognition==1.3.0
torch==1.3.1+cpu
selenium==3.141.0
gTTS==2.2.2
SoundFile==0.10.3.post1
docopt==0.6.2
pafy==0.5.5
librosa==0.8.0
requests==2.22.0
numpy==1.17.4
ffmpeg==1.4
SpeechRecognition==3.8.1
dlib==19.22.0
youtube_dl==2021.4.26
transformers==4.4.0
beautifulsoup4==4.9.3
google_api_python_client==2.10.0
imutils==0.5.4
ipython==7.25.0
path.py==12.5.0
Pillow==8.2.0
PySimpleGUI==4.45.0
scikit_learn==0.24.2
webdriver_manager==3.4.2
